A fix for #11 

Noticed some perf degradation after bumping the version of `next.jdbc`, could be some internal changes there:

1. [next.jdbc: porsas compiled](https://github.com/metosin/porsas/blob/b52322a9647b30419f79672c8f776daa6ad81e9e/test/porsas/core_test.clj#L91): from ~1200ns to ~1800ns
2. [next.jdbc](https://github.com/metosin/porsas/blob/b52322a9647b30419f79672c8f776daa6ad81e9e/test/porsas/core_test.clj#L96): from ~2400ns to ~3200ns